+++ src/controllers/UPC_controller.js
@@ -40,35 +40,5 @@ module.exports = {
         }catch(err){
             return res.status(400).send({error_comment_postage: err.message});
         }
-    },
-
-    async delete_postage_UPCs (req, res, next){
-        
-        try{
-            console.log("Removing Postage's UPCs...")
-
-            await UPC.deleteMany({ fk_postage_id: req.post._id })
-
-            console.log("Postage's UPCs successfully deleted!\n")
-
-            return next()
-        }catch(err){
-            return res.status(400).send({error_delete_postage_UPCs: err.message});
-        }
-    },
-
-    async delete_user_UPCs (req, res, next){
-        
-        try{
-            console.log("Removing User's UPCs...")
-
-            await UPC.deleteMany({ fk_user_id: req.params.id })
-
-            console.log("User's UPCs successfully deleted!\n")
-
-            return next()
-        }catch(err){
-            return res.status(400).send({error_delete_user_UPCs: err.message});
-        }
     }
 }
