+++ src/routes.js
@@ -18,7 +18,7 @@ router.get('/postage/list_all_with_UPS/:id', Postage.list_all_postages_with_UPS_
 router.delete('/postage/delete_for_test/:id', Postage.delete_one_for_test);
 router.delete('/postage/delete_all', Postage.delete_all);
 router.put('/postage/update_status/:id', Postage.update_status);
-router.put('/postage/delete_one', UPS.check_exist_user_and_postage, Postage.check_postage_is_not_anon, Postage.check_user_of_postage, UPS.delete_postage_UPSs, UPC.delete_postage_UPCs, Postage.delete_one);
+router.put('/postage/delete_one', UPS.check_exist_user_and_postage, Postage.check_postage_is_not_anon, Postage.check_user_of_postage, Postage.delete_postage_UPSs, Postage.delete_postage_UPCs, Postage.delete_one);
 router.put('/postage/update_one', UPS.check_exist_user_and_postage, Postage.check_postage_is_not_anon, Postage.check_user_of_postage, Postage.update_one);
 router.get('/postage/list_UPC/:id', Postage.list_UPCs_by_postage);
 
@@ -32,7 +32,7 @@ router.post('/user/login', Auth.authentication);
 router.get('/user/validate_session', Auth.session_authentication, Auth.find_user, Auth.refresh_token_and_data);
 router.get('/user/list_postages/:id', User.list_postages);
 router.delete('/user/delete_all', User.delete_all);
-router.delete('/user/delete/:id', Auth.find_user, UPS.delete_user_UPSs, UPC.delete_user_UPCs, User.delete_user_postages, User.delete);
+router.delete('/user/delete/:id', Auth.find_user, User.delete_user_UPSs, User.delete_user_UPCs, User.delete_user_postages, User.delete);
 router.put('/user/change_password/:id', Auth.find_user, User.change_password);
 
 //UPS routers
